+++ paymentsheet-example/src/androidTest/java/com/stripe/android/lpm/TestLinkCardBrand.kt
@@ -25,20 +25,10 @@ import com.stripe.android.utils.ForceNativeBankFlowTestRule
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import java.util.UUID
 
 @RunWith(AndroidJUnit4::class)
 internal class TestLinkCardBrand : BasePlaygroundTest() {
-
-    private val testParameters = TestParameters.create(
-        paymentMethodCode = "link",
-    ) { settings ->
-        settings[MerchantSettingsDefinition] = Merchant.US
-        settings[CurrencySettingsDefinition] = Currency.USD
-        settings[AutomaticPaymentMethodsSettingsDefinition] = false
-        settings[DefaultBillingAddressSettingsDefinition] = DefaultBillingAddress.On
-        settings[LinkSettingsDefinition] = true
-        settings[SupportedPaymentMethodsSettingsDefinition] = "card"
-    }
 
     @get:Rule
     val forceNativeBankFlowTestRule = ForceNativeBankFlowTestRule(
@@ -47,12 +37,12 @@ internal class TestLinkCardBrand : BasePlaygroundTest() {
 
     @Test
     fun testLinkCardBrandSuccess() {
-        val params = testParameters.copyPlaygroundSettings {
-            it[DefaultBillingAddressSettingsDefinition] = DefaultBillingAddress.On
-        }
+        val email = "email_${UUID.randomUUID()}@email.com"
+
+        testDriver.signUpForLink(makeSignUpTestParameters(email))
 
         testDriver.confirmLinkBankPayment(
-            testParameters = params,
+            testParameters = makeLinkTestParameters(email),
             afterAuthorization = { _, _ ->
                 rules.compose.waitUntil(DEFAULT_UI_TIMEOUT.inWholeMilliseconds) {
                     rules.compose
@@ -66,8 +56,12 @@ internal class TestLinkCardBrand : BasePlaygroundTest() {
 
     @Test
     fun testLinkCardBrandCancelAllowsUserToContinue() {
+        val email = "email_${UUID.randomUUID()}@email.com"
+
+        testDriver.signUpForLink(makeSignUpTestParameters(email))
+
         testDriver.confirmLinkBankPayment(
-            testParameters = testParameters.copy(
+            testParameters = makeLinkTestParameters(email).copy(
                 authorizationAction = AuthorizeAction.Cancel,
             ),
             afterAuthorization = { _, _ ->
@@ -75,5 +69,33 @@ internal class TestLinkCardBrand : BasePlaygroundTest() {
                     .waitFor(isEnabled())
             }
         )
+    }
+
+    private fun makeSignUpTestParameters(email: String): TestParameters {
+        return TestParameters.create(
+            paymentMethodCode = "card",
+            authorizationAction = null,
+            saveForFutureUseCheckboxVisible = true,
+        ) { settings ->
+            settings[MerchantSettingsDefinition] = Merchant.US
+            settings[CurrencySettingsDefinition] = Currency.USD
+            settings[AutomaticPaymentMethodsSettingsDefinition] = false
+            settings[DefaultBillingAddressSettingsDefinition] = DefaultBillingAddress.WithEmail(email)
+            settings[LinkSettingsDefinition] = true
+            settings[SupportedPaymentMethodsSettingsDefinition] = "card"
+        }
+    }
+
+    private fun makeLinkTestParameters(email: String): TestParameters {
+        return TestParameters.create(
+            paymentMethodCode = "link",
+        ) { settings ->
+            settings[MerchantSettingsDefinition] = Merchant.US
+            settings[CurrencySettingsDefinition] = Currency.USD
+            settings[AutomaticPaymentMethodsSettingsDefinition] = false
+            settings[DefaultBillingAddressSettingsDefinition] = DefaultBillingAddress.WithEmail(email)
+            settings[LinkSettingsDefinition] = true
+            settings[SupportedPaymentMethodsSettingsDefinition] = "card"
+        }
     }
 }