+++ paymentsheet/src/androidTest/java/com/stripe/android/paymentsheet/NetworkRuleExtensions.kt
@@ -1,6 +1,5 @@
 package com.stripe.android.paymentsheet
 
-import com.stripe.android.core.utils.urlEncode
 import com.stripe.android.networktesting.NetworkRule
 import com.stripe.android.networktesting.RequestMatcher
 import com.stripe.android.networktesting.RequestMatchers.host
@@ -16,7 +15,7 @@ internal fun NetworkRule.validateAnalyticsRequest(
         host("q.stripe.com"),
         method("GET"),
         query("event", eventName),
-        query("product_usage", urlEncode(productUsage.joinToString(","))),
+        query("product_usage", productUsage.joinToString(",")),
         *requestMatchers,
     ) { response ->
         response.status = "HTTP/1.1 200 OK"