+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/embedded/sheet/EmbeddedSheetComponent.kt
@@ -2,7 +2,7 @@ package com.stripe.android.paymentelement.embedded.sheet
 
 import android.app.Application
 import androidx.lifecycle.SavedStateHandle
-import com.stripe.android.common.di.ApplicationIdModule
+import com.stripe.android.common.di.ElementsSessionClientParamsModule
 import com.stripe.android.googlepaylauncher.injection.GooglePayLauncherModule
 import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
 import com.stripe.android.model.PaymentMethodCode
@@ -25,7 +25,7 @@ import javax.inject.Singleton
     modules = [
         EmbeddedActivityModule::class,
         EmbeddedCommonModule::class,
-        ApplicationIdModule::class,
+        ElementsSessionClientParamsModule::class,
         ExtendedPaymentElementConfirmationModule::class,
         GooglePayLauncherModule::class,
         EmbeddedLinkExtrasModule::class,