+++ paymentsheet/src/test/java/com/stripe/android/customersheet/CustomerSessionCustomerSheetActivityTest.kt
@@ -10,7 +10,6 @@ import androidx.lifecycle.testing.TestLifecycleOwner
 import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.stripe.android.core.utils.urlEncode
 import com.stripe.android.customersheet.util.CustomerSheetHacks
 import com.stripe.android.model.CardBrand
 import com.stripe.android.model.ElementsSession
@@ -477,11 +476,11 @@ class CustomerSessionCustomerSheetActivityTest {
     ) {
         networkRule.elementsSession(
             query("type", "deferred_intent"),
-            query(urlEncode("deferred_intent[setup_future_usage]"), "off_session"),
-            query(urlEncode("deferred_intent[mode]"), "setup"),
-            query(urlEncode("deferred_intent[payment_method_types][0]"), "card"),
-            query(urlEncode("deferred_intent[payment_method_types][1]"), "us_bank_account"),
-            query(urlEncode("deferred_intent[on_behalf_of]"), onBehalfOf),
+            query("deferred_intent[setup_future_usage]", "off_session"),
+            query("deferred_intent[mode]", "setup"),
+            query("deferred_intent[payment_method_types][0]", "card"),
+            query("deferred_intent[payment_method_types][1]", "us_bank_account"),
+            query("deferred_intent[on_behalf_of]", onBehalfOf),
             query("customer_session_client_secret", "cuss_123"),
         ) { response ->
             response.createElementsSessionResponse(